tests: Added a unit test for zero gradient PWLA, moved the tests into separate file for automatic discovery with pytest
